+++ CHANGELOG.md
@@ -9,4 +9,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Implemented - Deposit
 - Implemented - Withdrawal
 - Implemented - Transfer
+
 - Changed logic related to deposit flow
+- fixed user list item if user doesn't have username